Here is the file I could export from VBA
The original file format was .bas file that I then manually renamed to vbs. What you have access to in my repository for Mod_3_challenge is a txt file with my script on it  as well as a .vbs file. Our instructor stated pasting script into a notepad was fine.